requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.5.0
scipy>=1.7.0
scikit-image>=0.18.0
pillow>=8.3.0 